// web/src/app/api/generated/src/models/SetView.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Motion Ecosystem API
 * API for participant and specialist workflows
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { VersionSummary } from './VersionSummary';
import {
  VersionSummaryFromJSON,
  VersionSummaryFromJSONTyped,
  VersionSummaryToJSON,
  VersionSummaryToJSONTyped,
} from './VersionSummary';

/**
 *
 * @export
 * @interface SetView
 */
export interface SetView {
  /**
   *
   * @type {Date}
   * @memberof SetView
   */
  createdAt?: Date;
  /**
   *
   * @type {string}
   * @memberof SetView
   */
  id?: string;
  /**
   *
   * @type {string}
   * @memberof SetView
   */
  ownerAccountId?: string;
  /**
   *
   * @type {Array<VersionSummary>}
   * @memberof SetView
   */
  versions?: Array<VersionSummary>;
  /**
   *
   * @type {SetViewVisibilityEnum}
   * @memberof SetView
   */
  visibility?: SetViewVisibilityEnum;
}

/**
 * @export
 */
export const SetViewVisibilityEnum = {
  Private: 'PRIVATE',
  Shared: 'SHARED',
  Organization: 'ORGANIZATION',
} as const;
export type SetViewVisibilityEnum =
  (typeof SetViewVisibilityEnum)[keyof typeof SetViewVisibilityEnum];

/**
 * Check if a given object implements the SetView interface.
 */
export function instanceOfSetView(value: object): value is SetView {
  return true;
}

export function SetViewFromJSON(json: any): SetView {
  return SetViewFromJSONTyped(json, false);
}

export function SetViewFromJSONTyped(json: any, ignoreDiscriminator: boolean): SetView {
  if (json == null) {
    return json;
  }
  return {
    createdAt: json['createdAt'] == null ? undefined : new Date(json['createdAt']),
    id: json['id'] == null ? undefined : json['id'],
    ownerAccountId: json['ownerAccountId'] == null ? undefined : json['ownerAccountId'],
    versions:
      json['versions'] == null
        ? undefined
        : (json['versions'] as Array<any>).map(VersionSummaryFromJSON),
    visibility: json['visibility'] == null ? undefined : json['visibility'],
  };
}

export function SetViewToJSON(json: any): SetView {
  return SetViewToJSONTyped(json, false);
}

export function SetViewToJSONTyped(
  value?: SetView | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    createdAt: value['createdAt'] == null ? value['createdAt'] : value['createdAt'].toISOString(),
    id: value['id'],
    ownerAccountId: value['ownerAccountId'],
    versions:
      value['versions'] == null
        ? undefined
        : (value['versions'] as Array<any>).map(VersionSummaryToJSON),
    visibility: value['visibility'],
  };
}